Move interpolation routines to utils.py
