+++ roms_tools/setup/tides.py
@@ -6,7 +6,11 @@ from roms_tools.setup.grid import Grid
 from roms_tools.setup.plot import _plot
 from roms_tools.setup.fill import fill_and_interpolate
 from roms_tools.setup.datasets import Dataset
-from roms_tools.setup.utils import nan_check, interpolate_from_rho_to_u, interpolate_from_rho_to_v
+from roms_tools.setup.utils import (
+    nan_check,
+    interpolate_from_rho_to_u,
+    interpolate_from_rho_to_v,
+)
 from typing import Dict, List
 import matplotlib.pyplot as plt
